Switching to new project format and removed DIP related dependencies on VL.Core (moved in newer versions >=2019.1 to different assembly)
Also not that we need to re-define the external developer workflow when writing libraries for VL. I had to reference the needed assemblies inside VL.OpenCV.vl directly as the packages won't be present anymore in a final installation - shipping them with VL itself would blow up the download size from ~70mb to ~170mb which is just too much. Also our file selection process as already discussed often would not guarantee to do the same as what msbuild would've selected. So this commit will somehow break the external developer workflow (sorry randall, but as long as you don't update opencv you should be fine) until we manage to figure out a new way how to deal with it. Thinking about creating a separate VL.OpenCV.Demo project which is an exe project set to win-x64 and outputs to a local VL installation - that would enable nuget/msbuild to copy (and overwrite) the needed assemblies as well as native dlls.